src/DatabaseHelper/requirements.txt
redis
cassandra-driver
joblib
numpy
pandas
scikit-learn